Rename package to platform
