+++ pkg/platform/pullrequest.go
@@ -1,4 +1,18 @@
-package codereview
+// Copyright 2024 The Authors (see AUTHORS file)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package platform
 
 import (
 	"context"
@@ -7,7 +21,7 @@ import (
 	"github.com/abcxyz/guardian/pkg/github"
 )
 
-var _ CodeReview = (*PullRequest)(nil)
+var _ ChangeRequest = (*PullRequest)(nil)
 
 // PullRequestInput defines the required inputs for creating a new PullRequest
 // instance.
@@ -23,7 +37,7 @@ type PullRequestInput struct {
 	PullRequestNumber int
 }
 
-// PullRequest is the GitHub implementation for the CodeReview interface.
+// PullRequest is the GitHub implementation for the ChangeRequest interface.
 type PullRequest struct {
 	client github.GitHub
 	params *PullRequestInput
